pyproject.toml
[project]
name = "petmoji"
version = "0.1.0"
description = "반려동물 사진으로 AI 캐릭터 이모지 세트를 자동 생성하는 서비스"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["backend"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501", "E402", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
